pyproject.toml
[project]
name = "diagram-generator"
version = "0.1.0"
description = "A metadata-driven, FAANG-grade diagram generator following Hexagonal Architecture."
readme = "README.md"
authors = [
  {name = "Ali", email = "ali@google.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.7.0",
    "jinja2>=3.1.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "ruamel.yaml>=0.18.0",
    "lark>=1.1.9",
]

[project.optional-dependencies]
server = ["watchdog>=4.0.0"]
mcp = ["fastmcp>=0.1.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "datamodel-code-generator>=0.25.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
diagram-generator = "diagram_generator.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/diagram_generator"]


[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "datamodel-code-generator>=0.25.0",
    "types-pyyaml>=6.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF"]
ignore = []

[tool.mypy]
strict = true
python_version = "3.10"
ignore_missing_imports = true
